simple.ridge {ElemStatLearn}	R Documentation

Simple Ridge Regression

Description

Function to do simple ridge regression.

Usage

   simple.ridge(x, y, lambda = 1, df, ...)

Arguments

x
predictor matrix, NA's not allowed. Should not include a column of 1's.

y
response matrix, NA's not allowed

lambda
vector of ridge coefficients

df
vector of target degrees of freedom. lambda is calculated from these.

...
Presently not used

Details

Principal advantage with this compared to other implementations of ridge regression in R is that we can do the calculations for multiple degrees of freedom at once. Ridging is not applied to the intercept.

Value

List with components:

beta0
intercept

beta
matrix of coefficients. One column for each lambda

lambda
lambda used

df
df used

Author(s)

Kjetil B. Halvorsen

Examples

# See examples for \link{prostate}.

[Package ElemStatLearn version 2015.6.26.2 Index]