waveform {ElemStatLearn}	R Documentation

Function to simulate waveform data

Description

Simulates waveform data, as in the book page 402.

Usage

waveform(n)

Arguments

n
Number of waveforms to simulate

Details

The formulas are given in the book page 402.

Value

Dataframe with two variables, x, a matrix with 21 columns, and a vector y with class labels.

Author(s)

Kjetil Halvorsen

See Also

waveform.train and waveform.test.

[Package ElemStatLearn version 2015.6.26.2 Index]