zip2image {ElemStatLearn}	R Documentation

function to convert row of zip file to format used by image()

Description

This is a utility function converting zip.train/zip.test data to format useful for plotting with the function image.

Usage

zip2image(zip, line)

Arguments

zip
zip.train or zip.test.

line
row of matrix to take

Value

16 x 16 matrix suitable as argument for image.

Author(s)

Kjetil Halvorsen

Examples

## See example section of help file for zip.train

[Package ElemStatLearn version 2015.6.26.2 Index]